refactor(type): Extract loop-based date conversion into WideRangeDateConversion
