+++ velox/type/CMakeLists.txt
@@ -38,6 +38,7 @@ velox_add_library(
   TypeEncodingUtil.cpp
   TypeUtil.cpp
   Variant.cpp
+  WideRangeDateConversion.cpp
   HEADERS
   CastRegistry.h
   CastRule.h
@@ -61,6 +62,7 @@ velox_add_library(
   TypeEncodingUtil.h
   TypeUtil.h
   Variant.h
+  WideRangeDateConversion.h
 )
 
 velox_link_libraries(
